Reading Is Good
It is a backend service developed by using SpringBoot and MySQL. It runs on an embedded Tomcat via port 8181.
It would help if you created a Bearer token to send a request /authenticate endpoint, then used created Bearer token to send a request to other controller endpoints. Otherwise, an UnAuthorized exception is returned.